+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@ This release introduces the ability to send a flights report by email to the clu
 - Introduces the `flights_email_report_requested` event and the `FlightsEmailReportRequestedEventHandler`.
 - Introduces the `FlightsEmailReport` notification and the `FlightsEmailReportNotificationHandler`.
 - Implements the required methods to support the flights email report in the `I18n` class.
+- Adds a metadata section to the flights table that is sent by email.
 
 #### Frontend